+++ app/src/main/java/com/jpg/covidiando/ui/fraternity/FraternityFragment.java
@@ -1,0 +1,35 @@
+package com.jpg.covidiando.ui.fraternity;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.lifecycle.Observer;
+import androidx.lifecycle.ViewModelProviders;
+
+import com.jpg.covidiando.R;
+
+public class FraternityFragment extends Fragment {
+
+    private FraternityViewModel galleryViewModel;
+
+    public View onCreateView(@NonNull LayoutInflater inflater,
+                             ViewGroup container, Bundle savedInstanceState) {
+        galleryViewModel =
+                ViewModelProviders.of(this).get(FraternityViewModel.class);
+        View root = inflater.inflate(R.layout.fragment_fraternity, container, false);
+        //final TextView textView = root.findViewById(R.id.text_gallery);
+        galleryViewModel.getText().observe(getViewLifecycleOwner(), new Observer<String>() {
+            @Override
+            public void onChanged(@Nullable String s) {
+                //textView.setText(s);
+            }
+        });
+        return root;
+    }
+}
