+++ app/src/main/java/com/jpg/covidiando/ui/home/HomeFragment.java
@@ -4,22 +4,14 @@ import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.TextView;
 
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
-import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProviders;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.jpg.covidiando.item_list;
-import com.jpg.covidiando.recycler_adapter;
-import com.jpg.covidiando.ui.home.HomeViewModel;
 import com.jpg.covidiando.R;
-import com.jpg.covidiando.item_list;
-import com.jpg.covidiando.recycler_adapter;
 
 import java.util.ArrayList;
