--- app/src/main/java/com/jpg/covidiando/ui/shopping/ShoppingViewModel.java
@@ -1,19 +1,0 @@
-package com.jpg.covidiando.ui.shopping;
-
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
-import androidx.lifecycle.ViewModel;
-
-public class ShoppingViewModel extends ViewModel {
-
-    private MutableLiveData<String> mText;
-
-    public ShoppingViewModel() {
-        mText = new MutableLiveData<>();
-        mText.setValue("This is gallery fragment");
-    }
-
-    public LiveData<String> getText() {
-        return mText;
-    }
-}
