+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -31,9 +31,9 @@
 
     <fragment
         android:id="@+id/nav_gallery"
-        android:name="com.jpg.covidiando.ui.gallery.GalleryFragment"
-        android:label="@string/menu_gallery"
-        tools:layout="@layout/fragment_gallery" />
+        android:name="com.jpg.covidiando.ui.fraternityRed.FraternityRedViewModel"
+        android:label="Red de fraternidad"
+        tools:layout="@layout/fragment_fraternityred" />
 
     <fragment
         android:id="@+id/nav_slideshow"
